Reorganize the test suite to mirror the package layout
Tests for the generators were spread between the tests/ root (seven
files all exercising clientele/generators/api), a tests/generators/
root with misplaced files, and a tests/generators/framework/ directory
whose name matched nothing in the package. The tree now follows the
source tree, so the tests for a module are where you expect them:

- tests/generators/api/        APIGenerator and ClientsGenerator tests,
  including the feature/regression suites that lived at the tests/
  root (complex schemas, $ref handling, nullable fields, OpenAPI 3.1
  null types, issue 248, real-world fixture specs) and the former
  framework/ tests, renamed test_framework_*/test_standard_* ->
  test_api_generator_*
- tests/generators/basic/      BasicGenerator tests (split out of
  test_base_and_basic_coverage.py; its ClientsGenerator half went to
  tests/generators/api/test_clients.py)
- tests/generators/shared/     test_schemas.py (was
  tests/generators/test_schemas_coverage.py) next to test_utils.py
- tests/test_testing.py        renamed from test_testing_utilities.py
  to mirror clientele/testing.py

__init__.py files added to the new directories so the two
test_generator.py modules get unique import names.
