+++ tests/generators/api/test_openapi_31_null_type.py
