+++ tests/generators/api/test_ref_handling.py
